requirements.txt
Flask
SQLAlchemy
requests
python_jose